rename alibaba_cloud_slb to alibabacloudslb
