+++ pkg/controllers/alibabacloudslb/consts.go
@@ -14,6 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package alibaba_cloud_slb
+package alibabacloudslb
 
 const alibabaCloudSlbLbIdLabelKey = "service.k8s.alibaba/loadbalancer-id"
