+++ pkg/controllers/alibabacloudslb/types.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package alibaba_cloud_slb
+package alibabacloudslb
 
 import (
 	"fmt"
